Refactor instruction details components
